Version 0
This is a short program example, that show how to use the library uart.h. All functions of this library are not all use in this program. The purpose of this programme is to clear an UART terminal, display "Hello world!" at the cursor home position and go to the next line where the text sent to the MSP430 will be resend to the terminal.